--- src/constants/common/index.ts
@@ -1,1 +1,0 @@
-export * from './pagination';
