--- src/constants/common/pagination.ts
@@ -1,3 +1,0 @@
-export const FIRST_PAGE = 1;
-export const DEFAULT_PAGE_BUTTONS_SIZE = 10;
-export const DEFAULT_PAGING_SIZE = 20;
